Add option to export secrets as environment variables. (#289)
<!--
Thank you for proposing a pull request! Please note that SOME TESTS WILL
LIKELY FAIL due to how GitHub exposes secrets in Pull Requests from
forks.
Someone from the team will review your Pull Request and respond.

Please describe your change and any implementation details below.
-->

**This adds an option to export the secrets as environment variables.**

We need this, because we are migrating from using our own custom action
for fetching secrets and our old action would directly export the
secrets as environment variables, so to be able to maintain API
compatibility, we need this action to be able to do the same. Otherwise
we would have to make significant changes to our workflows to enable the
migration.

For others, this will be convenient, as most secrets you fetch will end
up in an environment variable anyway. This removes the need for the
manual step of putting the output from this action as an environment
variable in the step where it is used.

As per my implementation, the feature is disabled by default and needs
to be enabled to be used, which means that it is fully backwards
compatible and will not affect users that don't want to use this
feature.

I've tested it in one of our pipelines and it is working fine. I've also
run linting and tests locally, without any errors or failures.